# internal/app/controller/server/generateReport.go
package server

import (
	"encoding/csv"
	"encoding/json"
	"fmt"
	"github.com/darahayes/go-boom"
	_ "github.com/matster07/user-balance-service/docs" // docs is generated by Swag CLI, you have to import it.
	"github.com/matster07/user-balance-service/internal/pkg/logging"
	"net/http"
	"os"
	"path/filepath"
	"strconv"
)

//	@Summary      Generate CSV report
//	@Description  Генерация отчета услуг по прибыли в разрезе всех категорий услуг среди всех пользователей за месяц
//	@Tags         report
//	@Accept       json
//	@Produce      json
//  @Param        year  query uint true "Год"
//  @Param        month query uint true "Месяц"
//	@Router       /report/service/profit [get]
func (h *Handler) generateReport(w http.ResponseWriter, res *http.Request) {
	w.Header().Set("Content-Type", "application/json")

	year, err := strconv.ParseUint(res.URL.Query().Get("year"), 10, 16)
	month, err := strconv.ParseUint(res.URL.Query().Get("month"), 10, 16)

	filename := filepath.Join("reports", fmt.Sprintf("%d-%d-profit-report.csv", year, month))

	file, err := os.Create(filename)
	if err != nil {
		boom.BadRequest(w, "error while generating report")
		return
	}

	writer := csv.NewWriter(file)
	err = writer.Write([]string{"service_name", "profit"})
	if err != nil {
		boom.BadRequest(w, "error while generating report")
		return
	}

	rows, err := h.Order.GetDataForReport(uint(year), uint(month))
	if err != nil {
		boom.BadRequest(w, "error while generating report")
		return
	}

	h.populateZeroProfitServices(&rows)

	err = writer.WriteAll(rows)
	if err != nil {
		return
	}

	writer.Flush()

	err = json.NewEncoder(w).Encode(map[string]string{"stats": "success"})
	if err != nil {
		return
	}
}

func (h *Handler) populateZeroProfitServices(result *[][]string) {
	services, err := h.Service.FindAll()
	if err != nil {
		logging.GetLogger().Errorf("error while fetching services")
	}

	for _, value := range services {
		needToAdd := true

		for _, row := range *result {
			if row[0] == value.ServiceName {
				needToAdd = false
			}
		}

		if needToAdd {
			*result = append(*result, []string{value.ServiceName, "0"})
		}
	}
}